backend/requirements.txt
aioredis==1.3.1
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.2
Automat==0.8.0
boto3==1.12.5
botocore==1.15.5
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
coverage==5.0.3
cryptography==2.8
daphne==2.4.1
Django==3.0
django-cors-headers==3.2.0
django-eventstream==3.1.0
django-grip==3.0.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gripcontrol==4.0.0
hiredis==1.0.1
httplib2==0.17.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
isort==4.3.21
jmespath==0.9.4
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
Pillow==7.0.0
pluggy==0.13.1
psycopg2==2.8.4
pubcontrol==3.0.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytest==5.3.5
pytest-asyncio==0.10.0
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
selenium==3.141.0
service-identity==18.1.0
six==1.13.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
swapper==1.1.2.post1
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.1
wrapt==1.11.2
zipp==3.1.0
zope.interface==4.7.1
